# web-api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Web API Express",
        "description": "My WebApi using Node",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "tags": [
        {
            "name": "Players",
            "description": "API de jugadores en la aplicación"
        },
        {
            "name": "Avatars",
            "description": "API de avatars en la aplicación"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/getPlayers": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Obtiene todos los jugadores",
                "description": "Regresa todos los jugadores",
                "operationId": "getPlayers",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/PlayerAvatar"
                        }
                    }
                }
            }
        },
        "/api/getPlayer/{id}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Encuentra un jugador por id",
                "description": "Regresa el jugador con el id solicitado",
                "operationId": "getPlayer",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "id",
                        "description": "Id del jugador a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/PlayerAvatar"
                        }
                    }
                }
            }
        },
        "/api/addPlayer": {
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Agrega un nuevo jugador",
                "description": "Agrega un jugador a la base de datos",
                "operationId": "addPlayer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updatePlayer/{id}": {
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Actualiza un nuevo jugador",
                "description": "Actualiza un jugador a la base de datos",
                "operationId": "updatePlayer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "id",
                        "description": "Id del jugador a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevos jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    },
                    "400": {
                        "description": "Jugador Inválido"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }

                }
            }
        },
        "/api/deletePlayer/{id}": {
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Borra un jugador",
                "description": "Borra un jugador de la base de datos",
                "operationId": "deletePlayer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "id",
                        "description": "Id del jugador a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    },
                    "400": {
                        "description": "Jugador Inválido"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
        "/api/getAvatars": {
            "get": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Obtiene todos los avatars",
                "description": "Regresa todos los avatars",
                "operationId": "getAvatars",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Avatar"
                        }
                    }
                }
            }
        },
        "/api/getAvatar/{id}": {
            "get": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Encuentra un avatar por id",
                "description": "Regresa el jugador con el id solicitado",
                "operationId": "getAvatar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "id",
                        "description": "Id del avatar a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    }
                }
            }
        },
        "/api/addAvatar": {
            "post": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Agrega un nuevo avatar",
                "description": "Agrega un avatar a la base de datos",
                "operationId": "addAvatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo avatar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Avatar"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Avatar"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateAvatar/{id}": {
            "put": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Actualiza un avatar",
                "description": "Actualiza un avatar en la base de datos",
                "operationId": "updateAvatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "id",
                        "description": "Id del avatar a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevo avatar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Avatar"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Avatar"
                        }
                    },
                    "400": {
                        "description": "Avatar Inválido"
                    },
                    "404": {
                        "description": "Avatar no encontrado"
                    }

                }
            }
        },
        "/api/deleteAvatar/{id}": {
            "delete": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Borra un avatar",
                "description": "Borra un avatar de la base de datos",
                "operationId": "deleteAvatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "id",
                        "description": "Id del avatar a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Avatar"
                        }
                    },
                    "400": {
                        "description": "Avatar Inválido"
                    },
                    "404": {
                        "description": "Avatar no encontrado"
                    }
                }
            }
        }
    },
    "definitions": {
        "Player": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "playerName": {
                    "type": "string"
                },
                "avatarID": {
                    "type": "integer",
                    "format": "int64"
                },
                "points": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "Avatar": {
            "type": "object",
            "properties": {
                "avatarID": {
                    "type": "integer",
                    "format": "int64"
                },
                "avatarName": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                }
            }
        },
        "PlayerAvatar": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "playerName": {
                    "type": "string"
                },
                "avatarID": {
                    "type": "integer",
                    "format": "int64"
                },
                "points": {
                    "type": "integer",
                    "format": "int64"
                },
                "avatarName": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                }
            }
        }
    }
}